+++ .vscode/c_cpp_properties.json
@@ -1,22 +1,30 @@
 {
-    "configurations": [
-        {
-            "name": "Win32",
-            "includePath": [
-                "${workspaceFolder}/math/include/**/*",
-                "${workspaceFolder}/render/include/**/*",
-                "${workspaceFolder}/ui/include/**/*",
-                "${workspaceFolder}/utils/include/**/*",
-                "${workspaceFolder}/sample/cube/include/**/*"
-            ],
-            "defines": ["_DEBUG", "UNICODE", "_UNICODE"],
-            "windowsSdkVersion": "10.0.18362.0",
-            "cStandard": "c17",
-            "cppStandard": "c++20",
-            "intelliSenseMode": "windows-msvc-x64",
-            "compileCommands": "${workspaceFolder}/build/compile_commands.json",
-            "configurationProvider": "ms-vscode.cmake-tools"
-        }
-    ],
-    "version": 4
+  "configurations": [
+    {
+      "name": "Win32",
+      "includePath": [
+        "${workspaceFolder}/math/include/**/*",
+        "${workspaceFolder}/render/include/**/*",
+        "${workspaceFolder}/ui/include/**/*",
+        "${workspaceFolder}/utils/include/**/*",
+        "${workspaceFolder}/assimp/include/**/*",
+        "${workspaceFolder}/resources/include/**/*",
+        "${workspaceFolder}/scripting/include/**/*",
+        "${workspaceFolder}/webm/include/**/*",
+        "${workspaceFolder}/bass/include/**/*",
+        "${workspaceFolder}/engine/include/**/*",
+        "${workspaceFolder}/physics_2d/include/**/*",
+        "${workspaceFolder}/physics_3d/include/**/*",
+        "${workspaceFolder}/network/include/**/*"
+      ],
+      "defines": ["_DEBUG", "UNICODE", "_UNICODE"],
+      "windowsSdkVersion": "10.0.18362.0",
+      "cStandard": "c17",
+      "cppStandard": "c++20",
+      "intelliSenseMode": "windows-msvc-x64",
+      "compileCommands": "${workspaceFolder}/build/compile_commands.json",
+      "configurationProvider": "ms-vscode.cmake-tools"
+    }
+  ],
+  "version": 4
 }
