+++ math/src/easing.cpp
@@ -1,5 +1,7 @@
 #include <rawrbox/math/easing.hpp>
 
+#include <stdexcept>
+
 namespace rawrbox {
 	float EasingUtils::ease(rawrbox::Easing ease, float in) {
 		switch (ease) {
@@ -7,6 +9,8 @@ namespace rawrbox {
 				return in;
 			case Easing::STEP:
 				return in < 0.5F ? 0.F : 1.F;
+			default:
+				throw std::runtime_error("[RawrBox-Easing] Unsupported easing");
 		}
 	}
 } // namespace rawrbox
