+++ render/include/rawrbox/render/models/vertex.hpp
@@ -18,6 +18,19 @@ namespace rawrbox {
 		rawrbox::Vector4f uv = {};
 		rawrbox::Colorf color = {};
 
+		VertexData() = default;
+		explicit VertexData(const rawrbox::Vector3f& _pos,
+		    const rawrbox::Vector4f& _uv = {}, const rawrbox::Color cl = rawrbox::Colors::White()) : position(_pos), uv(_uv), color(cl) {}
+
+		explicit VertexData(const rawrbox::Vector3f& _pos,
+		    const rawrbox::Vector2f& _uv = {}, const rawrbox::Color cl = rawrbox::Colors::White()) : position(_pos), uv(_uv), color(cl) {}
+
+		// Atlas ---
+		void setAtlasId(uint32_t _id) {
+			this->uv.z = static_cast<float>(_id);
+		}
+		// ---------------------
+
 		static std::vector<Diligent::LayoutElement> vLayout(bool instanced = false) {
 			std::vector<Diligent::LayoutElement> v = {
 			    // Attribute 0 - Position
@@ -42,13 +55,16 @@ namespace rawrbox {
 	};
 
 	// Supports light ---
-	struct VertexNormData {
-		rawrbox::Vector3f position = {};
-		rawrbox::Vector4f uv = {};
-		rawrbox::Colorf color = {};
-
+	struct VertexNormData : public VertexData {
 		rawrbox::Vector3f normal = {};
 		rawrbox::Vector3f tangent = {};
+
+		VertexNormData() = default;
+		explicit VertexNormData(const rawrbox::Vector3f& _pos,
+		    const rawrbox::Vector4f& _uv = {}, const rawrbox::Vector3f& norm = {}, const rawrbox::Vector3f& tang = {}, const rawrbox::Color cl = rawrbox::Colors::White()) : rawrbox::VertexData(_pos, _uv, cl), normal(norm), tangent(tang) {}
+
+		explicit VertexNormData(const rawrbox::Vector3f& _pos,
+		    const rawrbox::Vector2f& _uv = {}, const rawrbox::Vector3f& norm = {}, const rawrbox::Vector3f& tang = {}, const rawrbox::Color cl = rawrbox::Colors::White()) : rawrbox::VertexData(_pos, _uv, cl), normal(norm), tangent(tang) {}
 
 		static std::vector<Diligent::LayoutElement> vLayout(bool instanced = false) {
 			std::vector<Diligent::LayoutElement> v = {
@@ -79,13 +95,44 @@ namespace rawrbox {
 	};
 
 	// Supports bones ---
-	struct VertexBoneData {
-		rawrbox::Vector3f position = {};
-		rawrbox::Vector4f uv = {};
-		rawrbox::Colorf color = {};
-
+	struct VertexBoneData : public VertexData {
 		std::array<uint32_t, rawrbox::MAX_BONES_PER_VERTEX> bone_indices = {};
 		std::array<float, rawrbox::MAX_BONES_PER_VERTEX> bone_weights = {};
+
+		VertexBoneData() = default;
+		explicit VertexBoneData(const rawrbox::Vector3f& _pos,
+		    const rawrbox::Vector4f& _uv = {}, const rawrbox::Color cl = rawrbox::Colors::White()) : rawrbox::VertexData(_pos, _uv, cl) {}
+
+		explicit VertexBoneData(const rawrbox::Vector3f& _pos,
+		    const rawrbox::Vector2f& _uv = {}, const rawrbox::Color cl = rawrbox::Colors::White()) : rawrbox::VertexData(_pos, _uv, cl) {}
+
+		// BONES UTILS -----
+		int index = 0;
+		void addBoneData(uint32_t boneId, float weight) {
+			if (index < rawrbox::MAX_BONES_PER_VERTEX) {
+				this->bone_indices[index] = boneId;
+				this->bone_weights[index] = weight;
+
+				index++;
+			} else {
+				// find the bone with the smallest weight
+				int minIndex = 0;
+				float minWeight = this->bone_weights[0];
+				for (int i = 1; i < rawrbox::MAX_BONES_PER_VERTEX; i++) {
+					if (this->bone_weights[i] < minWeight) {
+						minIndex = i;
+						minWeight = this->bone_weights[i];
+					}
+				}
+
+				// replace with new bone if the new bone has greater weight
+				if (weight > minWeight) {
+					this->bone_indices[minIndex] = boneId;
+					this->bone_weights[minIndex] = weight;
+				}
+			}
+		}
+		// ------------------
 
 		static std::vector<Diligent::LayoutElement> vLayout(bool instanced = false) {
 			std::vector<Diligent::LayoutElement> v = {
@@ -115,16 +162,44 @@ namespace rawrbox {
 	};
 
 	// Supports light && bones ---
-	struct VertexNormBoneData {
-		rawrbox::Vector3f position = {};
-		rawrbox::Vector4f uv = {};
-		rawrbox::Colorf color = {};
-
-		rawrbox::Vector3f normal = {};
-		rawrbox::Vector3f tangent = {};
-
+	struct VertexNormBoneData : public VertexNormData {
 		std::array<uint32_t, rawrbox::MAX_BONES_PER_VERTEX> bone_indices = {};
 		std::array<float, rawrbox::MAX_BONES_PER_VERTEX> bone_weights = {};
+
+		VertexNormBoneData() = default;
+		explicit VertexNormBoneData(const rawrbox::Vector3f& _pos,
+		    const rawrbox::Vector4f& _uv = {}, const rawrbox::Vector3f& norm = {}, const rawrbox::Vector3f& tang = {}, const rawrbox::Color cl = rawrbox::Colors::White()) : rawrbox::VertexNormData(_pos, _uv, norm, tang, cl) {}
+
+		explicit VertexNormBoneData(const rawrbox::Vector3f& _pos,
+		    const rawrbox::Vector2f& _uv = {}, const rawrbox::Vector3f& norm = {}, const rawrbox::Vector3f& tang = {}, const rawrbox::Color cl = rawrbox::Colors::White()) : rawrbox::VertexNormData(_pos, _uv, norm, tang, cl) {}
+
+		// BONES UTILS -----
+		int index = 0;
+		void addBoneData(uint32_t boneId, float weight) {
+			if (index < rawrbox::MAX_BONES_PER_VERTEX) {
+				this->bone_indices[index] = boneId;
+				this->bone_weights[index] = weight;
+
+				index++;
+			} else {
+				// find the bone with the smallest weight
+				int minIndex = 0;
+				float minWeight = this->bone_weights[0];
+				for (int i = 1; i < rawrbox::MAX_BONES_PER_VERTEX; i++) {
+					if (this->bone_weights[i] < minWeight) {
+						minIndex = i;
+						minWeight = this->bone_weights[i];
+					}
+				}
+
+				// replace with new bone if the new bone has greater weight
+				if (weight > minWeight) {
+					this->bone_indices[minIndex] = boneId;
+					this->bone_weights[minIndex] = weight;
+				}
+			}
+		}
+		// ------------------
 
 		static std::vector<Diligent::LayoutElement> vLayout(bool instanced = false) {
 			std::vector<Diligent::LayoutElement> v = {
@@ -157,95 +232,82 @@ namespace rawrbox {
 		}
 	};
 
-	// Supports light && bones ---
-	// With helpers, but not to be passed to buffer
-	struct ModelVertexData : public VertexNormBoneData {
-		// BLEND HELPERS ---
-		rawrbox::Vector3f ori_pos = {};
-		rawrbox::Vector3f ori_norm = {};
-		// -----------------
+	/*
+		// Supports light && bones ---
+		// With helpers, but not to be passed to buffer
+		struct ModelVertexData : public VertexNormBoneData {
+			// BLEND HELPERS ---
+			rawrbox::Vector3f ori_pos = {};
+			rawrbox::Vector3f ori_norm = {};
+			// -----------------
 
-		using VertexNormBoneData::vLayout;
+			using VertexNormBoneData::vLayout;
 
-		ModelVertexData() = default;
-		explicit ModelVertexData(const rawrbox::Vector3f& _pos,
-		    const rawrbox::Vector4f& _uv = {}, const rawrbox::Vector3f& norm = {}, const rawrbox::Vector3f& tang = {}, const rawrbox::Color cl = rawrbox::Colors::White()) : ori_norm(norm) {
-			this->position = _pos;
-			this->uv = _uv;
-			this->normal = norm;
-			this->tangent = tang;
-			this->color = cl;
-		}
-
-		explicit ModelVertexData(const rawrbox::Vector3f& _pos,
-		    const rawrbox::Vector2f& _uv = {}, const rawrbox::Vector3f& norm = {}, const rawrbox::Vector3f& tang = {}, const rawrbox::Color cl = rawrbox::Colors::White()) : ori_pos(_pos), ori_norm(norm) {
-			this->position = _pos;
-			this->uv = _uv;
-			this->normal = norm;
-			this->tangent = tang;
-			this->color = cl;
-		}
-
-		explicit ModelVertexData(const rawrbox::Vector3f& _pos,
-		    const rawrbox::Vector4f& _uv = {}, const rawrbox::Color cl = rawrbox::Colors::White()) : ori_pos(_pos) {
-			this->position = _pos;
-			this->uv = _uv;
-			this->color = cl;
-		}
-
-		explicit ModelVertexData(const rawrbox::Vector3f& _pos,
-		    const rawrbox::Vector2f& _uv = {}, const rawrbox::Color cl = rawrbox::Colors::White()) : ori_pos(_pos) {
-			this->position = _pos;
-			this->uv = _uv;
-			this->color = cl;
-		}
-
-		// GPU Picker ---
-		void setId(uint32_t _id) {
-			// if (id > 0x00FFFFFF) throw std::runtime_error("[RawrBox-Mesh] Invalid id");
-			// this->id = 0xFF000000 | _id; // Alpha is not supported;
-		}
-		// ---------------------
-
-		// Atlas ---
-		void setAtlasId(uint32_t _id) {
-			this->uv.z = static_cast<float>(_id);
-		}
-		// ---------------------
-
-		// BONES UTILS -----
-		int index = 0;
-		void addBoneData(uint32_t boneId, float weight) {
-			if (index < rawrbox::MAX_BONES_PER_VERTEX) {
-				this->bone_indices[index] = boneId;
-				this->bone_weights[index] = weight;
-
-				index++;
-			} else {
-				// find the bone with the smallest weight
-				int minIndex = 0;
-				float minWeight = this->bone_weights[0];
-				for (int i = 1; i < rawrbox::MAX_BONES_PER_VERTEX; i++) {
-					if (this->bone_weights[i] < minWeight) {
-						minIndex = i;
-						minWeight = this->bone_weights[i];
-					}
-				}
-
-				// replace with new bone if the new bone has greater weight
-				if (weight > minWeight) {
-					this->bone_indices[minIndex] = boneId;
-					this->bone_weights[minIndex] = weight;
-				}
+			ModelVertexData() = default;
+			explicit ModelVertexData(const rawrbox::Vector3f& _pos,
+			    const rawrbox::Vector4f& _uv = {}, const rawrbox::Vector3f& norm = {}, const rawrbox::Vector3f& tang = {}, const rawrbox::Color cl = rawrbox::Colors::White()) : ori_norm(norm) {
+				this->position = _pos;
+				this->uv = _uv;
+				this->normal = norm;
+				this->tangent = tang;
+				this->color = cl;
 			}
-		}
-		// ------------------
 
-		// BLEND SHAPE UTILS ---
-		void reset() {
-			this->position = this->ori_pos;
-			this->normal = this->ori_norm;
-		}
-		// ---------------------
-	};
+			explicit ModelVertexData(const rawrbox::Vector3f& _pos,
+			    const rawrbox::Vector2f& _uv = {}, const rawrbox::Vector3f& norm = {}, const rawrbox::Vector3f& tang = {}, const rawrbox::Color cl = rawrbox::Colors::White()) : ori_pos(_pos), ori_norm(norm) {
+				this->position = _pos;
+				this->uv = _uv;
+				this->normal = norm;
+				this->tangent = tang;
+				this->color = cl;
+			}
+
+			explicit ModelVertexData(const rawrbox::Vector3f& _pos,
+			    const rawrbox::Vector4f& _uv = {}, const rawrbox::Color cl = rawrbox::Colors::White()) : ori_pos(_pos) {
+				this->position = _pos;
+				this->uv = _uv;
+				this->color = cl;
+			}
+
+			explicit ModelVertexData(const rawrbox::Vector3f& _pos,
+			    const rawrbox::Vector2f& _uv = {}, const rawrbox::Color cl = rawrbox::Colors::White()) : ori_pos(_pos) {
+				this->position = _pos;
+				this->uv = _uv;
+				this->color = cl;
+			}
+
+			// GPU Picker ---
+			void setId(uint32_t _id) {
+				// if (id > 0x00FFFFFF) throw std::runtime_error("[RawrBox-Mesh] Invalid id");
+				// this->id = 0xFF000000 | _id; // Alpha is not supported;
+			}
+			// ---------------------
+
+			// Atlas ---
+			void setAtlasId(uint32_t _id) {
+				this->uv.z = static_cast<float>(_id);
+			}
+			// ---------------------
+
+
+
+			// BLEND SHAPE UTILS ---
+			void reset() {
+				this->position = this->ori_pos;
+				this->normal = this->ori_norm;
+			}
+			// ---------------------
+		};
+		*/
+
+	// UTILS ---
+	template <typename T>
+	concept noSupport = requires(T t) { !t.bone_indices && !t.normal; };
+
+	template <typename T>
+	concept supportsBones = requires(T t) { t.bone_indices; };
+
+	template <typename T>
+	concept supportsNormals = requires(T t) { t.normal; };
+	// ---
 } // namespace rawrbox
