+++ samples/001-stencil/src/game.cpp
@@ -100,7 +100,7 @@ namespace stencil {
 		// ----
 
 		// Textures ---
-		auto mesh = rawrbox::MeshUtils::generateCube({0, 0, 0}, {2.F, 2.F, 2.F});
+		auto mesh = rawrbox::MeshUtils::generateCube<>({0, 0, 0}, {2.F, 2.F, 2.F});
 		mesh.setTexture(this->_texture3);
 
 		std::random_device prng;
