+++ samples/003-assimp/include/assimp/game.hpp
@@ -2,21 +2,22 @@
 
 #include <rawrbox/assimp/model.hpp>
 #include <rawrbox/engine/engine.hpp>
+#include <rawrbox/render/materials/skinned.hpp>
 #include <rawrbox/render/models/text3D.hpp>
 
 #include <memory>
 
 namespace assimp {
 	class Game : public rawrbox::Engine {
-		std::unique_ptr<rawrbox::AssimpModel> _model = std::make_unique<rawrbox::AssimpModel>();
-		std::unique_ptr<rawrbox::AssimpModel> _model2 = std::make_unique<rawrbox::AssimpModel>();
-		std::unique_ptr<rawrbox::AssimpModel> _model3 = std::make_unique<rawrbox::AssimpModel>();
-		std::unique_ptr<rawrbox::AssimpModel> _model4 = std::make_unique<rawrbox::AssimpModel>();
-		std::unique_ptr<rawrbox::AssimpModel> _model5 = std::make_unique<rawrbox::AssimpModel>();
-		std::unique_ptr<rawrbox::AssimpModel> _model6 = std::make_unique<rawrbox::AssimpModel>();
+		std::unique_ptr<rawrbox::AssimpModel<>> _model = std::make_unique<rawrbox::AssimpModel<>>();
+		std::unique_ptr<rawrbox::AssimpModel<>> _model2 = std::make_unique<rawrbox::AssimpModel<>>();
+		std::unique_ptr<rawrbox::AssimpModel<rawrbox::MaterialSkinned>> _model3 = std::make_unique<rawrbox::AssimpModel<rawrbox::MaterialSkinned>>();
+		std::unique_ptr<rawrbox::AssimpModel<rawrbox::MaterialSkinned>> _model4 = std::make_unique<rawrbox::AssimpModel<rawrbox::MaterialSkinned>>();
+		std::unique_ptr<rawrbox::AssimpModel<rawrbox::MaterialSkinned>> _model5 = std::make_unique<rawrbox::AssimpModel<rawrbox::MaterialSkinned>>();
+		std::unique_ptr<rawrbox::AssimpModel<>> _model6 = std::make_unique<rawrbox::AssimpModel<>>();
 
-		std::unique_ptr<rawrbox::Model> _modelGrid = std::make_unique<rawrbox::Model>();
-		std::unique_ptr<rawrbox::Text3D> _text = std::make_unique<rawrbox::Text3D>();
+		std::unique_ptr<rawrbox::Model<>> _modelGrid = std::make_unique<rawrbox::Model<>>();
+		std::unique_ptr<rawrbox::Text3D<>> _text = std::make_unique<rawrbox::Text3D<>>();
 
 		rawrbox::Font* _font = nullptr;
